surian voice agent/requirements.txt
streamlit
fastapi
uvicorn
openai-whisper
elevenlabs
requests
python-dotenv
pydub
soundfile
livekit 